requirements.txt
pandas==2.0.3
plotly==5.9.0
PyYAML==6.0
PyYAML==6.0.1
streamlit==1.32.2
streamlit_authenticator==0.3.1
streamlit_option_menu==0.3.12
supabase==2.4.0
